dav: replace fileid in name with detectionid
This allows you to unambiguously move detections from one cluster
to another instead of files. As a result, if two detections from the
same file end up in one cluster, we don't have any funny behavior.

Note that it is perfectly possible to have multiple detections for
the same file in the same cluster, e.g. when a person is present
multiple times inside a single image. In this case, we need to do
de-duplication somewhere, but this is currently unimplemented anyway.

Signed-off-by: Varun Patil <varunpatil@ucla.edu>